RC 3.0 (18). Update Launch Image with version. Change name to 'Locos d'Azotea' Fix Quote+.swift for previous versions. AboutViewController: pdfView dependes of iOS 11.

+++ ios/LDLARadio/MVC-MVVM/Model/Quote+.swift
@@ -60,7 +60,7 @@ extension Quote {
         if n == 0 {
             return nil
         }
-        let rnd = Int(arc4random())%n
+        let rnd = Int(arc4random()%UInt32(n))
         return array?[rnd]
     }
